[Task] Ajustes no BI

+++ src/main/java/school/sptech/conmusicapi/modules/user/mapper/UserMapper.java
@@ -56,8 +56,8 @@ public class UserMapper {
     ) {
         UserKpiDto kpiDto = new UserKpiDto();
 
-        double percentageConfirmed = (double) (confirmed / negotiations) * 100;
-        double percentageCanceled = (double) (confirmed / negotiations) * 100;
+        double percentageConfirmed =  ((double) confirmed / (double) negotiations) * 100;
+        double percentageCanceled = ((double) confirmed / (double) negotiations) * 100;
 
         kpiDto.setReceivedProposals(receivedProposals);
         kpiDto.setNegotiations(negotiations);
